+++ src/main/java/com/develonity/comment/repository/CommentRepository.java
@@ -28,5 +28,5 @@ public interface CommentRepository extends JpaRepository<Comment, Long> {
 
   boolean existsCommentsByBoardId(Long boardId);
 
-  int countByBoardId(Long boardId);
+  long countByBoardId(Long boardId);
 }
